Move useState and useErrorState to other file for coverage.
