+++ viewCommon/src/main/java/com/tfandkusu/template/viewmodel/UnidirectionalViewModel.kt
@@ -1,7 +1,5 @@
 package com.tfandkusu.template.viewmodel
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.livedata.observeAsState
 import androidx.lifecycle.LiveData
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
@@ -21,11 +19,4 @@ interface UnidirectionalViewModel<EVENT, EFFECT, STATE> {
     fun createEffectChannel(): Channel<EFFECT> {
         return Channel(Channel.UNLIMITED)
     }
-}
-
-@Composable
-inline fun <reified STATE, EFFECT, EVENT> useState(
-    viewModel: UnidirectionalViewModel<EVENT, EFFECT, STATE>
-): STATE {
-    return viewModel.state.observeAsState(viewModel.createDefaultState()).value
 }
