+++ viewCommon/src/main/java/com/tfandkusu/template/viewmodel/UseState.kt
@@ -1,0 +1,11 @@
+package com.tfandkusu.template.viewmodel
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.livedata.observeAsState
+
+@Composable
+inline fun <reified STATE, EFFECT, EVENT> useState(
+    viewModel: UnidirectionalViewModel<EVENT, EFFECT, STATE>
+): STATE {
+    return viewModel.state.observeAsState(viewModel.createDefaultState()).value
+}
