+++ viewCommon/src/main/java/com/tfandkusu/template/viewmodel/error/UseErrorState.kt
@@ -1,0 +1,9 @@
+package com.tfandkusu.template.viewmodel.error
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.livedata.observeAsState
+
+@Composable
+fun useErrorState(apiErrorViewModelHelper: ApiErrorViewModelHelper): ApiErrorState {
+    return apiErrorViewModelHelper.state.observeAsState(ApiErrorState()).value
+}
